filetree: move files into better modules
