+++ src/repositoryConnection/TemplateManager.ts
