+++ src/views/PublishStatusBar.ts
